refactor: repository가 dao를 생성자로 받도록 변경

+++ app/src/main/java/woowacourse/shopping/ui/cart/CartProductAdapter.kt
@@ -6,13 +6,13 @@ import woowacourse.shopping.model.CartProduct
 
 class CartProductAdapter(
     items: List<CartProduct>,
-    onClickDelete: (position: Int) -> Unit,
+    onClickDelete: (position: Long) -> Unit,
     private val dateFormatter: DateFormatter,
 ) : RecyclerView.Adapter<CartProductViewHolder>() {
 
     private val items: MutableList<CartProduct> = items.toMutableList()
 
-    private val onClickDelete = { position: Int ->
+    private val onClickDelete = { position: Long ->
         onClickDelete(position)
         removeItem(position)
     }
@@ -27,8 +27,8 @@ class CartProductAdapter(
 
     override fun getItemCount(): Int = items.size
 
-    private fun removeItem(position: Int) {
-        items.removeAt(position)
-        notifyItemRemoved(position)
+    private fun removeItem(position: Long) {
+        items.removeAt(position.toInt())
+        notifyItemRemoved(position.toInt())
     }
 }
